+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [tbd] date
+### Added
+- Raise error on if `status code >= 400` (similar to `phoenix_test 0.6`, although error page)
+
 ## [0.6.0] 2025-03-18
 ### Added
 - Add and remove cookies: `add_cookies/2`, `add_session_cookie/3`, `clear_cookies/{1,2}` @peaceful-james